Update install instructions for swanctl vs ipsec
When we released version 1.9.40 of the ipsec-release bosh addon, we
changed the internal command used to create the secure network. We
changed from using the `ipsec` command, to using the `swanctl` command
(both of which are provided by the `strongswan` software package).

As a result, the command that users should use to check if their
installation is working has also changed. Previously it was `ipsec
statusall`. Now it is either or both of `swanctl --stats` or `swanctl
--list-conns`.

This change updates the docs to assume we're using >= 1.9.40, and
instruct the user to use `swanctl`. For users who find themselves
reading recent docs while debugging older systems (or scripts that were
written for older systems), we retain a note about the old way of doing
things.

[#184475207]